perf: drop translations proxy
Since translation files are now served from the `apps` domain directly.
We are still seeing some low number of requests towards the translations domain, but we already deployed the the frontend that is no longer using it a week ago.
Users affected by this will either not notice or just need to reload the page.

https://bettermarks.atlassian.net/browse/OT-1